% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldinGroundOz-package.R
\docType{package}
\name{GoldinGroundOz}
\alias{GoldinGroundOz-package}
\alias{GoldinGroundOz}
\title{Exploring Australia's Simulated Housing Market Dynamics}
\description{
The GoldinGroundOz package demonstrates the long-term movements of
Australia's detached house prices, household disposable income,
and average capital gain rate changes, using simulated datasets.
}
\section{Datasets}{

\itemize{
\item \code{housepr_income}: yearly simulated index of detached house prices and household disposable income.
\item \code{capital_gain}: simulated average capital gain rate changes under different policy phases.
}
}

\section{Shiny Application}{

The package includes an interactive Shiny app launched by \code{run_shinyapp()}
for visual exploration of the simulated data.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://Acrea-QwQ.github.io/GoldinGroundOz}
  \item \url{https://etc5523-2025.github.io/assignment-4-packages-and-shiny-apps-Acrea-QwQ/}
}

}
\author{
\strong{Maintainer}: Xingyu Wang \email{xwan0510@student.monash.edu}

}
